requirements.txt
sqlalchemy
mysql-connector-python
pymysql
python-dotenv
pytubefix
psutil